vshelp: use sfile to guess the script's location
Trick I found in speeddating: sfile gives the file being sourced
(vshelp.vim in this case), so we can trim off the filename and plugin
directory to get either the bundle dir or the vim root (or something
else crazy). So long as it's not too crazy, this should work.

Don't expose vshelp_devenv. Switch from global to script-local variable.

Fail if we can't find the script. We can't do much without it.