Add trick to convert registers to pointers
Add Visual Studio trick to make it easier to guess pointers from the
register window. Useful when debugging Release builds where Visual
Studio won't do it for you itself. Result is a list of possible
candidates (currently only works for x86_64).

Made it an autoload function with no command because it's infrequently
used.