Handle filenames containing spaces
Vim escapes the filenames but that confuses the python API. Strip
escaping and use python's path module to clean up the path.

Should probably do this all in python instead. (Would handle
noshellslash.)

I think this was the cause of the unhelpful pywintypes.com_error. The
only other thing I've seen is trying to open a nonexistant file. Remove
it for now.